+++ tests/t-utf8.c
@@ -77,7 +77,7 @@ test_open(void)
 	BOOL ok = CreateDirectoryW(wpath, NULL);
 	if (!ok) {
 		DWORD e = GetLastError();
-		wprintf(L"Cannot create directory %ls (code %u)\n", wpath, e);
+		wprintf(L"Cannot create directory %ls (code %lu)\n", wpath, (unsigned long int)e);
 		abort();
 	}
